# [nexmark] Reduce Send/Recv with batched receivers